+++ CHANGELOG.md
@@ -5,6 +5,15 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
+### Fixed
+- Mask subdomains in email addresses correctly
+### Changed
+- Renamed `fromName` (`FROM_NAME`) to `emailSignature` (`EMAIL_SIGNATURE`)
+- Use `display_name` personnel property in email templates.
+### Removed
+- Removed ReCAPTCHA from front-end config data
+
+## [4.0.0] - 2019-04-11
 ### Added
 - Integrated Personnel\IdBroker component library code
 - Integrated PasswordStore components library code
@@ -81,7 +90,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ### Added
 - Initial version of Password Manager Backend.
 
-[Unreleased]: https://github.com/silinternational/idp-pw-api/compare/3.0.0...HEAD
+[Unreleased]: https://github.com/silinternational/idp-pw-api/compare/4.0.0...HEAD
+[3.0.0]: https://github.com/silinternational/idp-pw-api/compare/3.0.0...4.0.0
 [3.0.0]: https://github.com/silinternational/idp-pw-api/compare/2.2.0...3.0.0
 [2.2.0]: https://github.com/silinternational/idp-pw-api/compare/2.1.4...2.2.0
 [2.1.4]: https://github.com/silinternational/idp-pw-api/compare/2.1.3...2.1.4